@@ -24,9 +24,9 @@
     },
     "require-dev": {
         "payum/core": "^1.5@dev",
-        "phpunit/phpunit": "~4.0",
+        "phpunit/phpunit": "^5.7",
         "php-http/guzzle6-adapter": "^1.0",
-        "symfony/phpunit-bridge": "~2.8|~3.0"
+        "symfony/phpunit-bridge": "^3.1|^4"
     },
     "config": {
         "bin-dir": "bin"
