@@ -7,7 +7,7 @@ use Payum\Core\HttpClientInterface;
 use Payum\Paypal\Ipn\Api;
 use Psr\Http\Message\RequestInterface;
 
-class ApiTest extends \PHPUnit_Framework_TestCase
+class ApiTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @test
@@ -185,7 +185,7 @@ class ApiTest extends \PHPUnit_Framework_TestCase
      */
     protected function createHttpClientMock()
     {
-        return $this->getMock('Payum\Core\HttpClientInterface', array('send'));
+        return $this->createMock('Payum\Core\HttpClientInterface', array('send'));
     }
 
     /**
